+++ Sources/ZcashLightClientKit/Synchronizer/CombineSDKSynchronizer.swift
@@ -309,6 +309,16 @@ extension CombineSDKSynchronizer: CombineSynchronizer {
         }
     }
 
+    public func httpGetOverTor(
+        for request: URLRequest,
+        retryLimit: UInt8,
+        timeoutMilliseconds: UInt64
+    ) -> SinglePublisher<(data: Data, response: HTTPURLResponse), Error> {
+        AsyncToCombineGateway.executeThrowingAction() {
+            try await self.synchronizer.httpGetOverTor(for: request, retryLimit: retryLimit, timeoutMilliseconds: timeoutMilliseconds)
+        }
+    }
+
     public var broadcaster: Broadcaster { synchronizer.broadcaster }
 
     public func rewind(_ policy: RewindPolicy) -> CompletablePublisher<Error> { synchronizer.rewind(policy) }